+++ src/vmobjects/VMInteger.cpp
@@ -31,7 +31,7 @@ VMInteger::VMInteger() {
 	embeddedInteger = 0;
 }
 
-VMInteger::VMInteger(int32_t val) {
+VMInteger::VMInteger(long val) {
 	embeddedInteger = val;
 }
